server/requirements.txt
langchain-neo4j
neo4j
langchain-openai
langchain-community
langchain-experimental
unstructured
markdown
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
langchain
openai
python-dotenv
bcrypt==4.0.1
python-jose[cryptography] 
passlib
python-multipart
apscheduler
qdrant-client>=1.7.0
pgvector
fastapi-mail
alembic
requests
langchain-text-splitters
pytest==8.3.4
pytest-asyncio==0.24.0
httpx
psycopg2-binary
